Add provisioning checks to start saga unwind test
The tests still pass (the provisioning problems in this saga have to do with
side effects from a failure in the ensure_running node) but these are still good
checks to have.
