+++ nexus/src/app/sagas/test_helpers.rs
@@ -11,7 +11,9 @@ use crate::{
     Nexus,
 };
 use async_bb8_diesel::{AsyncRunQueryDsl, AsyncSimpleConnection};
-use diesel::{ExpressionMethods, QueryDsl, SelectableHelper};
+use diesel::{
+    BoolExpressionMethods, ExpressionMethods, QueryDsl, SelectableHelper,
+};
 use futures::future::BoxFuture;
 use nexus_db_queries::{
     authz,
@@ -184,6 +186,68 @@ pub async fn instance_fetch(
               "instance_and_vmm" => ?db_state);
 
     db_state
+}
+
+pub async fn no_virtual_provisioning_resource_records_exist(
+    cptestctx: &ControlPlaneTestContext,
+) -> bool {
+    use nexus_db_queries::db::model::VirtualProvisioningResource;
+    use nexus_db_queries::db::schema::virtual_provisioning_resource::dsl;
+
+    let datastore = cptestctx.server.apictx().nexus.datastore().clone();
+    let conn = datastore.pool_connection_for_tests().await.unwrap();
+
+    datastore
+        .transaction_retry_wrapper("no_virtual_provisioning_resource_records_exist")
+        .transaction(&conn, |conn| async move {
+            conn
+                .batch_execute_async(nexus_test_utils::db::ALLOW_FULL_TABLE_SCAN_SQL)
+                .await
+                .unwrap();
+
+            Ok(
+                dsl::virtual_provisioning_resource
+                    .filter(dsl::resource_type.eq(nexus_db_queries::db::model::ResourceTypeProvisioned::Instance.to_string()))
+                    .select(VirtualProvisioningResource::as_select())
+                    .get_results_async::<VirtualProvisioningResource>(&conn)
+                    .await
+                    .unwrap()
+                    .is_empty()
+            )
+        }).await.unwrap()
+}
+
+pub async fn no_virtual_provisioning_collection_records_using_instances(
+    cptestctx: &ControlPlaneTestContext,
+) -> bool {
+    use nexus_db_queries::db::model::VirtualProvisioningCollection;
+    use nexus_db_queries::db::schema::virtual_provisioning_collection::dsl;
+
+    let datastore = cptestctx.server.apictx().nexus.datastore().clone();
+    let conn = datastore.pool_connection_for_tests().await.unwrap();
+
+    datastore
+        .transaction_retry_wrapper(
+            "no_virtual_provisioning_collection_records_using_instances",
+        )
+        .transaction(&conn, |conn| async move {
+            conn.batch_execute_async(
+                nexus_test_utils::db::ALLOW_FULL_TABLE_SCAN_SQL,
+            )
+            .await
+            .unwrap();
+            Ok(dsl::virtual_provisioning_collection
+                .filter(
+                    dsl::cpus_provisioned.ne(0).or(dsl::ram_provisioned.ne(0)),
+                )
+                .select(VirtualProvisioningCollection::as_select())
+                .get_results_async::<VirtualProvisioningCollection>(&conn)
+                .await
+                .unwrap()
+                .is_empty())
+        })
+        .await
+        .unwrap()
 }
 
 /// Tests that the saga described by `dag` succeeds if each of its nodes is
